[OWL] Refactor Shortest path code in order to implement several algorithms
